requirements.txt
aioredis==1.3.1
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.1.0
autobahn==24.4.2
Automat==24.8.1
beautifulsoup4==4.13.3
billiard==4.2.1
bootstrap4==0.1.0
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
channels==4.2.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==44.0.1
daphne==4.1.2
Django==5.1.6
django-bootstrap4==24.4
djangorestframework==3.15.2
h11==0.14.0
hiredis==3.1.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
kombu==5.4.2
msgpack==1.1.0
pillow==11.1.0
prompt_toolkit==3.0.50
psycopg==3.2.4
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyOpenSSL==25.0.0
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
redis==5.2.1
service-identity==24.2.0
setuptools==75.8.0
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
vine==5.1.0
wcwidth==0.2.13
websockets==14.2
zope.interface==7.2
pytest
pytest-django